deleted BookingFunctions and BookingRoutes and combined into BookingController.js, createBooking and calculateTotalPrice function adjustment, booking routes set in server.js
